Begin Moving Errors to quick-error
