+++ diesel/Cargo.toml
@@ -17,6 +17,7 @@ byteorder = "0.3.*"
 quickcheck = { version = "0.3.1", optional = true }
 chrono = { version = "^0.2.17", optional = true }
 uuid = { version = ">=0.2.0, <0.4.0", optional = true, features = ["use_std"] }
+quick-error = "1.1.0"
 
 [dev-dependencies]
 quickcheck = "0.3.1"
